Fix long pressed accents on Chrome+macOS
Fix copied from https://github.com/ianstormtaylor/slate/pull/3113

On macOS, long press triggers an IME to insert an accent. Browsers that implements beforeinput event uses the insertReplacementText and a proper range. Chrome is instead selecting the character to replace, right when the input is triggered.

If we detect macOS && Chrome, the we proceed to check if the DOM selection has been updated before the input event and a detection of a selection change. The check is very naive but enough for this particular case. This can't happen when previous editor selection was expanded I guess.

I had to expose through a query the composition state of the before plugin, this is used to discard completely this fix in case of a user using another IME.

Co-Authored-By: Matthew Saunders <8563678+msaun008@users.noreply.github.com>