Avoid extra `select` events
React sends a synthetic `select` event after certain events
(like `keydown`) if the selection has changed. This can cause onSelect
handlers to run in the same event loop, after the value has been
modified, but before the component has rerendered.

When this happens, the browser value is out of sync with the editor
value, so browser selections may no longer apply. Additionally, the
selection set by the browser overrides the selection set by Slate,
putting the cursor in the wrong position.

If Slate changes the document, it knows better than the browser where
the selection would go, so it should ignore the `select` event.