ACT-24: Fix input events in Chromium-based browsers
Slate marks the editor as plaintext-only, in order to avoid showing
the formatting toolbar on mobile and touch devices. These built-in
formatting options cause the editor to get out of sync, so we don't
want them to appear. This also disallows built-in formatting shortcut
keys that we might not handle, which would also break the editor.

Chrome, Edge and other Chromium-based browsers treat the
plaintext-only editor as a textarea instead of a contenteditable for
purposes of input events. This means that they don't send us the
target range for replacement events (things like spellcheck, inserting
accented characters, etc.), because the spec says not to for textareas
and input elements. Without the target range, we can't know which part
of the editor to replace with the replaced word, and it doesn't seem
possible to infer from the information on the event.

Safari sends the target range for these events even if the editor says
it's plaintext-only, which is why it works fine.

The best fix seems to be to remove the plaintext-only attribute from
the contenteditable and ignore formatting input events. Slate can't
handle them because it doesn't know the editor schema. It might be
possible for a plugin to override onBeforeInput to correctly handle
those events in a specific type of editor.