CREATE-621: Normalize inserts at the edges of nodes
When reacting to the insert text input event, Slate inserts the
content using the range given by the browser. This allows IME events
and other text replacements to work correctly.

The browser's selection is not always consistent with value.selection
-- for example, the browser may target the last position inside an
element, while value.selection targets the first position outside the
element.

It is more natural to insert at value.selection during this ambiguous
situation, so value.selection should override the browser's selection
here. This also gives the editor full control over the insert point,
which is necessary for sticky inlines and similar plugins.